% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_to.R
\name{class_to}
\alias{class_to}
\title{class_to}
\usage{
class_to(x, new_class)
}
\arguments{
\item{x}{object}

\item{new_class}{new class to convert to}
}
\description{
convert object to new class
}
\keyword{internal}
