% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html2tagList.R
\name{html2tagList}
\alias{html2tagList}
\title{html2tagList}
\usage{
html2tagList(x)
}
\arguments{
\item{x}{character vector of html}
}
\description{
convert raw html to tagList
}
\examples{
\donttest{
x<-'<h1>Title</h1>
   <h2>Header text</h2>
   <p>Text here</p>
   <h1>Title</h1>
   <h2>Header text</h2>
   <p>Text here</p>'
   
 html2tagList(x)  

require(xtable)
htmlIn<-print(xtable::xtable(mtcars),type = 'html',print.results = F)
htmlIn
tagL<-html2tagList(htmlIn)
class(tagL)
tagL
if(interactive()) htmltools::browsable(tagL)
}
}
\keyword{internal}
