% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSwitch.R
\name{importSwitch}
\alias{importSwitch}
\title{importSwitch}
\usage{
importSwitch(file)
}
\arguments{
\item{file}{path to imported file}
}
\value{
data.frame
}
\description{
utility function to read imported data
}
\keyword{internal}
