% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_heatmaply.R
\name{launch_heatmaply}
\alias{launch_heatmaply}
\title{launch_heatmaply}
\usage{
launch_heatmaply(obj, minHeight = 1000)
}
\arguments{
\item{obj}{data.frame or list of data.frames}

\item{minHeight}{numeric that gets passed to paneViewer}
}
\description{
launch shiny app UI for heatmaply from R console
}
\examples{
if(interactive()){
data(mtcars)
launch_heatmaply(mtcars)

data(iris)
launch_heatmaply(list('Example1'=mtcars,'Example2'=iris))
}
}
