% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_heatmaply.R
\name{launch_heatmaply}
\alias{launch_heatmaply}
\title{launch_heatmaply}
\usage{
launch_heatmaply(obj, plotHeight = 800, viewerType = "paneViewer")
}
\arguments{
\item{obj}{data.frame or list of data.frames}

\item{plotHeight}{numeric that sets the height of the plot output (default 800px)}

\item{viewerType}{character of the viewer to be used to launch the app to c('paneViewer','dialogViewer','browserViewer')}
}
\description{
launch shiny app UI for heatmaply from R console
}
\examples{
\dontrun{
if(interactive()){
data(mtcars)
launch_heatmaply(mtcars)

data(iris)
launch_heatmaply(list('Example1'=mtcars,'Example2'=iris))
}
}
}
